Rename Java package to com.farpost.groovyshell
Align the Java package name with the new Maven coordinates
(com.farpost:groovy-shell). Previously the source still lived under
me.bazhenov.groovysh, which was confusing for consumers who see
com.farpost in their POM but com/farpost-unrelated imports in their code.

- Move all sources from me/bazhenov/groovysh/ to com/farpost/groovyshell/
- Update package declarations and intra-package imports
- Update README Spring bean XML and mvn exec:java examples

Note: this is a breaking change for any consumer of 2.2.4, who must
update their imports in addition to the groupId change. Given that the
groupId change already forced a POM update, the additional import
rename is a small marginal cost for a much cleaner consumer experience.

Verified end-to-end in a Spring Boot 2.7 smoke test on both JDK 8 and
JDK 21: app boots, GroovyShellService binds port 6789, SSH banner is
served, clean shutdown via @PreDestroy.
