+++ groovy-shell/src/main/java/com/farpost/groovyshell/GroovyShellService.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package me.bazhenov.groovysh;
+package com.farpost.groovyshell;
 
 import org.apache.groovy.groovysh.Groovysh;
 import org.apache.sshd.common.session.Session;
