+++ groovy-shell/src/main/java/com/farpost/groovyshell/Main.java
@@ -1,4 +1,4 @@
-package me.bazhenov.groovysh;
+package com.farpost.groovyshell;
 
 import java.io.IOException;
